requirements.txt
flask
flask-cors
torch
torchvision
torchaudio
torch_geometric
rdkit-pypi
py3Dmol
requests
wikipedia
pandas
numpy
scikit-learn
networkx
